C4-50 start bar_plot_chart rewrite
